Client and Admin View Payment
Changed the structure of the database to include the Payment collection to be a sub collection in the Project collection. So that admin can issue more than 1 payment for a project. The admin can view the payment list and the client can view the latest payment in the project list page.